clase inicio
para separar metodos y quede mas organizado
